Model training properties (#47)
Training properties are added - launch_training_completion and launch_training_chat for commands for starting training the models.
Also properties lora_completion and lora_chat were added for the location of the .gguf lora files. If not empty - options for lora adapters are added on starting the server from properties launch_completion and launch_chat.
The chat with UI for now can't use Lora adapter. This will require a change in webui.